+++ src/QiParsers.h
@@ -51,6 +51,8 @@ inline bool parseNumber(char decimalMark, char groupingMark, Iterator& first,
 
   double sum = 0, denom = 1;
   NumberState state = STATE_INIT;
+  bool seenNumber = false;
+  double sign = 1.0;
 
   Iterator cur = first;
   for(; cur != last; ++cur) {
@@ -58,9 +60,11 @@ inline bool parseNumber(char decimalMark, char groupingMark, Iterator& first,
     case STATE_INIT:
       if (*cur == '-') {
         state = STATE_LHS;
+        sign = -1.0;
       } else if (*cur == decimalMark) {
         state = STATE_RHS;
       } else if (*cur >= '0' && *cur <= '9') {
+        seenNumber = true;
         state = STATE_LHS;
         sum = *cur - '0';
       } else {
@@ -73,6 +77,7 @@ inline bool parseNumber(char decimalMark, char groupingMark, Iterator& first,
       } else if (*cur == decimalMark) {
         state = STATE_RHS;
       } else if (*cur >= '0' && *cur <= '9') {
+        seenNumber = true;
         sum *= 10;
         sum += *cur - '0';
       } else if (*cur == '-') {
@@ -88,6 +93,7 @@ inline bool parseNumber(char decimalMark, char groupingMark, Iterator& first,
       } else if (*cur == decimalMark) {
         return false;
       } else if (*cur >= '0' && *cur <= '9') {
+        seenNumber = true;
         denom *= 10;
         sum += (*cur - '0') / denom;
       } else {
@@ -96,15 +102,18 @@ inline bool parseNumber(char decimalMark, char groupingMark, Iterator& first,
       break;
     case STATE_FIN:
       last = cur++;
-      res = sum;
-      return true;
+      res = sign * sum;
+      return seenNumber;
     }
   }
 
   // Hit the end of the string, so must be done
   last = cur;
-  res = sum;
-  return true;
+  res = sign * sum;
+
+  // Only true if we saw a number and reached the end of the string without
+  // finishing the number
+  return seenNumber && state != STATE_FIN;
 }